+++ compiler/root.zig
@@ -2,6 +2,7 @@ pub const AST = @import("AST.zig");
 pub const Compilation = @import("Compilation.zig");
 pub const Diagnostic = @import("Diagnostic.zig");
 pub const Source = @import("Source.zig");
+pub const backend = @import("backend/root.zig");
 pub const spell = @import("util/spell.zig");
 
 test {